refactor: move Registration page according to the new project structure
