+++ src/app/features/auth/registration/registration.component.html
@@ -1,1 +1,120 @@
-<p>registration works!</p>
+<!-- eslint-disable @angular-eslint/template/cyclomatic-complexity -->
+<form [formGroup]="registrationForm" (ngSubmit)="onSubmit()">
+  <div class="field">
+    <label for="email">Email</label>
+    <div class="input-container">
+      <input id="email" formControlName="email" />
+      <div *ngIf="registrationForm.get('email')?.invalid && registrationForm.get('email')?.touched">
+        <span class="error-line" *ngIf="getFieldError('email') === 'required'">{{ `email ${errMsg.REQUIRED}` }}</span>
+        <span class="error-line" *ngIf="getFieldError('email') === 'pattern'">{{ `${errMsg.INVALID} email format` }}</span>
+      </div>
+    </div>
+  </div>
+
+  <div class="field">
+    <label for="password">Password</label>
+    <div class="input-container">
+      <input id="password" [type]="passwordFieldType" formControlName="password" />
+      <button type="button" (click)="togglePasswordVisibility()">{{ isPasswordVisible ? '🙈' : '👁️‍🗨️' }}</button>
+      <div *ngIf="registrationForm.get('password')?.invalid && registrationForm.get('password')?.touched">
+        <span class="error-line" *ngIf="getFieldError('password') === 'required'">{{ `password ${errMsg.REQUIRED}` }}</span>
+        <span class="error-line" *ngIf="getFieldError('password') === 'pattern'">{{ `password ${errMsg.PASSWORD}` }}</span>
+      </div>
+    </div>
+  </div>
+
+  <div class="field">
+    <label for="firstName">First Name</label>
+    <div class="input-container">
+      <input id="firstName" formControlName="firstName" />
+      <div *ngIf="registrationForm.get('firstName')?.invalid && registrationForm.get('firstName')?.touched">
+        <span class="error-line" *ngIf="getFieldError('firstName') === 'required'">{{ `first name ${errMsg.REQUIRED}` }}</span>
+        <span class="error-line" *ngIf="getFieldError('firstName') === 'pattern'">{{ `first name ${errMsg.NAME}` }}</span>
+      </div>
+    </div>
+  </div>
+
+  <div class="field">
+    <label for="lastName">Last Name</label>
+    <div class="input-container">
+      <input id="lastName" formControlName="lastName" />
+      <div *ngIf="registrationForm.get('lastName')?.invalid && registrationForm.get('lastName')?.touched">
+        <span class="error-line" *ngIf="getFieldError('lastName') === 'required'">{{ `last name ${errMsg.REQUIRED}` }}</span>
+        <span class="error-line" *ngIf="getFieldError('lastName') === 'pattern'">{{ `last name ${errMsg.NAME}` }}</span>
+      </div>
+    </div>
+  </div>
+
+  <div class="field">
+    <label for="dob">Date of Birth</label>
+    <div class="input-container input-container_bod">
+      <input id="dob" type="date" formControlName="dob" />
+      <div *ngIf="registrationForm.get('dob')?.invalid && registrationForm.get('dob')?.touched">
+        <span class="error-line" *ngIf="getFieldError('dob') === 'required'">{{ `date of birth ${errMsg.REQUIRED}` }}</span>
+        <span class="error-line" *ngIf="getFieldError('dob') === 'dob'">{{ registrationForm.get('dob')?.errors?.['dob'].message }}</span>
+      </div>
+    </div>
+  </div>
+
+  <div formGroupName="address">
+    <h3>Address</h3>
+    <div class="field">
+      <label for="street">Street</label>
+      <div class="input-container">
+        <input id="street" formControlName="street" />
+        <div
+          *ngIf="registrationForm.controls['address'].get('street')?.invalid && registrationForm.controls['address'].get('street')?.touched"
+        >
+          <span class="error-line" *ngIf="getFieldError('street') === 'required'">{{ `street ${errMsg.REQUIRED}`
+          }}</span>
+        </div>
+      </div>
+    </div>
+
+    <div class="field">
+      <label for="city">City</label>
+      <div class="input-container">
+        <input id="city" formControlName="city" />
+        <div *ngIf="registrationForm.controls['address'].get('city')?.invalid && registrationForm.controls['address'].get('city')?.touched">
+          <span class="error-line" *ngIf="getFieldError('city') === 'required'">{{ `city ${errMsg.REQUIRED}` }}</span>
+          <span class="error-line" *ngIf="getFieldError('city') === 'pattern'">{{ `city ${errMsg.NAME}` }}</span>
+        </div>
+      </div>
+    </div>
+
+    <div class="field">
+      <label for="postalCode">Postal Code</label>
+      <div class="input-container">
+        <input id="postalCode" formControlName="postalCode" />
+        <div
+          *ngIf="
+            registrationForm.controls['address'].get('postalCode')?.invalid &&
+            registrationForm.controls['address'].get('postalCode')?.touched
+          "
+        >
+          <span class="error-line" *ngIf="getFieldError('postalCode') === 'required'">{{ `postal code ${errMsg.REQUIRED}` }}</span>
+          <span class="error-line" *ngIf="getFieldError('postalCode') === 'message'">{{ getPostalCodeErrorMessage() }}</span>
+        </div>
+      </div>
+    </div>
+
+    <div class="field">
+      <label for="country">Country</label>
+      <div class="input-container">
+        <select id="country" formControlName="country">
+          <option value="" disabled>Select a country</option>
+          <option *ngFor="let country of validCountries" [value]="country">{{ country }}</option>
+        </select>
+        <div
+          *ngIf="
+            registrationForm.controls['address'].get('country')?.invalid && registrationForm.controls['address'].get('country')?.touched
+          "
+        >
+          <span class="error-line" *ngIf="getFieldError('country') === 'required'">{{ `country ${errMsg.REQUIRED}` }}</span>
+        </div>
+      </div>
+    </div>
+  </div>
+
+  <button type="submit" [disabled]="registrationForm.invalid">Register</button>
+</form>
